Add files via upload
With the first-ever ban on social media handle, it is clear that the abuse of information rights and the improper application of power are reaching a new level in Nigeria, meriting widespread concern. This study aimed to examine, in retrospect, the quality of newspaper framing of the twitter ban issue in Nigeria, where the media’s efficacy in covering recent events in the country has been questioned due to persistent public opinion that the media is performing its fourth estate of the realm’s duties in accordance to odious governing standards. This study examined the newspaper framing of President Mohammadu Buhari’s ban on twitter in Nigeria with particular emphasis on a The Punch and Daily Trust newspapers. The framing theory of the mass media was adopted in the study and the content analysis research design was as well employed. The finding of the study indicated among other things, that both ‘Daily Trust’ and ‘The Punch’ newspapers framed reports of twitter ban in Cause, Effect and Reaction frames. The study concluded that both newspapers considered in the study had covered the issue of the Nigerian twitter ban in different measures and perspectives, as 
arising from the differences in the slants accorded the reportage. The study recommended that the government should uphold the tenet of a true democratic state by ensuring that no ban is placed on social networking sites as such actions undermines the fundamental human right of expression of the citizens.